desktop: split config dialogs by flow
